Refactor BrowserWarning into page and component
 Due to the need to show the browser warning on patients route when use has data storage set up we needed to break out the content of the page into a component and then add a conditional in the patients render flow.

 Styling has also been applied to conform with standard blip layout of pages
