+++ app/components/browserwarning/chrome-logo.png
