+++ app/components/browserwarning/index.js
@@ -1,0 +1,3 @@
+import BrowserWarning from  './browserwarning';
+
+export default BrowserWarning;
