+++ app/style.less
@@ -32,6 +32,7 @@
 @import "components/waitlist/waitlist.less";
 @import "components/chart/settings.less";
 @import "components/chart/modalsubnav.less";
+@import "components/browserwarning/browserwarning.less";
 
 @import "pages/app/app.less";
 @import "pages/login/login.less";
@@ -45,5 +46,5 @@
 @import "pages/patientdata/patientdata.less";
 @import "pages/passwordreset/passwordreset.less";
 @import "pages/emailverification/emailverification.less";
-@import "pages/browserwarning/browserwarning.less";
 @import "pages/terms/terms.less";
+@import "pages/browserwarning/browserwarning.less";
