+++ test/unit/components/browserwarning.test.js
@@ -7,7 +7,7 @@ var React = require('react');
 var TestUtils = require('react-addons-test-utils');
 var expect = chai.expect;
 
-var BrowserWarning = require('../../../app/pages/browserwarning');
+var BrowserWarning = require('../../../app/components/browserwarning');
 
 describe('BrowserWarning', function () {
